+++ packages/fern-docs/syntax-highlighter/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./CodeBlockWithClipboardButton";
 export * from "./FernSyntaxHighlighter";
 export type { ScrollToHandle } from "./FernSyntaxHighlighterTokens";
-export { SyntaxHighlighterFeatureFlagsProvider } from "./SyntaxHighlighterFeatureFlags";
+export { SyntaxHighlighterEdgeFlagsProvider } from "./SyntaxHighlighterEdgeFlags";
